Co-locate connection schema to match previous test.
